Fix streaming chunks
Despite being told that this was a bug in Excon that was fixed (and
later became effective in the `docker-api` gem), the `:chunk_size`
option on Docker's connection must be set to 1 to stream output
normally. Elsewise, Excon just buffers output from a connection before
finally flushing it. For more context, see swipely/docker-api#77

This fixes #13

Signed-off-by: David Celis <me@davidcel.is>